requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
dj-static==0.0.6
django-extensions=3.1.5
Django==4.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.5
sqlparse==0.4.2
static3==0.7.0
django-test-without-migrations=0.6